% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptron.R
\name{perceptron}
\alias{perceptron}
\title{Square a number
works with 'iris' data set}
\usage{
perceptron(x)
}
\arguments{
\item{x}{A data set}
}
\value{
returns squared number
}
\description{
Square a number
works with 'iris' data set
}
